backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-docx==1.1.0
PyPDF2==3.0.1
openai==1.3.0
pydantic==2.5.0
python-dotenv==1.0.0
email-validator==2.1.0
firebase-admin==7.1.0
mangum==0.17.0
boto3==1.34.0
